+++ example/lib/login_view_model.dart
@@ -4,14 +4,15 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:uniform/uniform.dart';
+import 'package:uniform_example/login_page.dart';
 import 'package:uniform_example/validators/email_input_field_validator.dart';
 import 'package:uniform_example/validators/password_input_field_validator.dart';
 
 enum LoginFormTags {
   email,
   password,
+  selectGender,
   gender,
-  rememberMe,
 }
 
 class LoginViewModel extends ChangeNotifier {
@@ -19,14 +20,50 @@ class LoginViewModel extends ChangeNotifier {
     formController = FormController(
       validators: {const InputFieldValidator.required()},
     );
-    _initValidators();
+
+    _emailField = TextFieldController.create(
+      formController,
+      tag: LoginFormTags.email,
+    )..setValidators({const EmailInputFieldValidator()});
+
+    _passwordField = TextFieldController.create(
+      formController,
+      tag: LoginFormTags.password,
+      autoValidate: true,
+    )..setValidators({const PasswordInputFieldValidator()});
+
+    FieldController<bool>.create(
+      formController,
+      tag: LoginFormTags.selectGender,
+    )
+      ..setInitialValue(false)
+      ..onUpdate(_onSelectGenderUpdate);
+
+    _genderField = FieldController<Gender>.create(
+      formController,
+      tag: LoginFormTags.gender,
+    );
+
+    formController.setInitialValues(
+      {
+        LoginFormTags.email: 'sales@acme-software.com',
+        LoginFormTags.gender: Gender.female,
+      },
+    );
   }
 
   late final FormController formController;
 
+  late final TextFieldController _emailField;
+  late final TextFieldController _passwordField;
+  late final FieldController<Gender> _genderField;
+
   Map<Object, Object?> _userData = {};
+  bool _requireGender = false;
 
   Map<Object, Object?> get userData => _userData;
+
+  bool get requireGender => _requireGender;
 
   Future<void> login() async {
     if (formController.validate()) {
@@ -35,20 +72,21 @@ class LoginViewModel extends ChangeNotifier {
       // Simulates login
       await Future<void>.delayed(const Duration(seconds: 2));
 
-      _userData = formController.values;
+      _userData = {
+        'Email Address': _emailField.value,
+        'Password': _passwordField.value,
+        'Gender': _genderField.value?.name,
+      };
       notifyListeners();
 
       formController.setSubmitted(false);
     }
   }
 
-  Future<void> _initValidators() async {
-    final emailController = await formController(LoginFormTags.email);
-    emailController
-      ..setValidators({const EmailInputFieldValidator()})
-      ..setValue('sales@acme-software.com');
+  void _onSelectGenderUpdate(bool? selectGender) {
+    _requireGender = selectGender ?? false;
+    notifyListeners();
 
-    final passwordController = await formController(LoginFormTags.password);
-    passwordController.setValidators({const PasswordInputFieldValidator()});
+    if (!_requireGender) formController.deactivate(LoginFormTags.gender);
   }
 }